requirements.txt
backcall==0.2.0
decorator==4.4.2
gitdb==4.0.7
gitpython==3.1.14
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
numpy==1.19.5
pandas==1.1.5
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.8.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
smmap==4.0.0
traitlets==4.3.3
wcwidth==0.2.5
